# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for KDS backend.",
        "title": "KDS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8630",
    "basePath": "/api",
    "paths": {
...
        "/oauth/google/redirect": {
            "get": {
                "description": "Handle Google OAuth redirect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GoogleLoginResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.GoogleLoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}